Identifying a reliable cross section

Identifies a "representative" cross section for a bullet land engraved area. Striation marks on a bullet land are the best expressed at the heel (bottom) of a bullet where break-off is still problematic. Using cross-correlation we identify a cross section that is the closest to the bottom of the bullet but does not suffer from break-off. If the resulting cross section is equal to the maximum of the search area (defined in xlimits), there should be some investigation, whether this cross section is usable. There is the risk of tank rash. XXX still to do: are missing values only on the right hand side (leading shoulder)?

bulletCheckCrossCut(path, bullet = NULL, distance = 25, xlimits = c(50,
  500), minccf = 0.9, span = 0.03, percent_missing = 50)

Arguments

path
path to an x3p file. Ignored in case bullet is not NULL.

bullet
if not NULL, the bullet land engraved area (in x3p format).

distance
positive numeric value indicating the distance between cross sections to use for a comparison

xlimits
vector of values between which to check for cross sections in a stable region

minccf
minimal value of cross correlation to indicate a stable region

span
The span for the loess smooth function

percent_missing
maximum percent missing values on the crosscut to be picked